execution/state: fold code-hash test and bench into code_read_parallel_test.go
